Created utils.py --> Added initialization and cloning parameter functions; Updated PSO.py, GA.py --> imported functions from utils
