+++ optimizers/GA.py
@@ -1,0 +1,5 @@
+from typing import Dict, List, Callable, Type, Iterable, Optional
+import torch # type: ignore
+from torch.optim import Optimizer # type: ignore
+from .utils import _initialize_param_groups, clone_param_group, clone_param_groups
+
